Ugly hack using reflection to fix long espresso tests.
Since espresso 3.x, the `RootViewPicker` class waits until the root view in question becomes available, with a timeout backoff. We have tests that expect to NOT find a root view, and every time this check was being performed, it took over 60 seconds for espresso to finally give up and tell us that the root view wasn't found.

The workaround is to use reflection on the `WindowManager` class to count the number of root views.