Prevent some random test failures.
When an instrumentation test started after a previous test completed, a shared prefs change was triggered for the poem content.  There was no actual change (the saved poem was null before and after), but the shared preference change triggered a live data change, which the `ReaderFragment` observed, invalidating the options menu.  If the test was in the process of typing a query, this caused the test to fail, as the search view was reinitialized.