Replace AsyncTask/Handler threading with Kotlin coroutines.
As for the idling resource for the coroutine threading:

For some reason, even though we unregister the resource at the end of each test, it was still being used.
The fix is to make sure the name of the resource is unique for each test.
Hope this won't cause memory leaks :/