Fix process-text intent filter.
Poet assistant stops appearing in the text selection menu in some Google apps (chrome, keep, ...), starting in Android 14.

This weird change, to add the `<intent-filter>` for the `VIEW` action and the `https` data schema, seems to make this work again.

Also: add the `android:label` to the `<activity>` elements to make the proper labels appear in some apps which were just showing "Poet Assistant" (ex: Firefox).