Replace the workaround in tests that was pulling down the app bar.
Instead of having this workaround in tests: prevent the issue in the first place.

It was being caused by the query word, in the results list header, being selectable. This was making the app bar scroll up just when opening the app, even if there were no queries yet.

Only make this `TextView` selectable if it's from an actual user query.