Migrate dagger to hilt.
Migrate dagger components and modules to hilt.

Add a NonAndroidEntryPoint for dependencies which need to be accessed from a non-android (non-hilt) class.

Replace DaggerHelper methods with injection via this NonAndrydEntryPoint.

Adjust the PoetAssistantApplication initialization: set the theme as soon as the settings are accessed, instead of from the Application.onCreate() method.

Remove "retain fragment" from the ReaderFragment. This was causing a crash with hilt: "java.lang.IllegalStateException: onAttach called multiple times with different Context! Hilt Fragments should not be retained."

For tests:
* Add @HiltAndroidTest.
* Add the HiltAndroidRule.
* For androidTests:
  - Use a specific test runner that specifies a hilt application class.
* For tests (robolectric):
  - Clear the static AndroidViewModelFactory singleton instance, which was introducing failures as the application instance from a test 1 was being reused in a test 2.